move the remaining files from the reducers to the store dir
