+++ src/store/reducer.js
@@ -2,10 +2,10 @@ import { combineReducers } from '@reduxjs/toolkit';
 import { reducer as formReducer } from 'redux-form';
 
 import ui from './ui';
-import app from '../store/app';
+import app from './app';
 import messages from './messages';
-import modals from '../store/modals';
-import channels from '../store/channels';
+import modals from './modals';
+import channels from './channels';
 
 export default combineReducers({
   channels: channels.reducer,
